修复phoneNumber越界问题，重新建表

+++ demo/hotelBooking/models.py
@@ -2,10 +2,9 @@ from django.db import models
 from django.contrib.auth.models import User
 # Create your models here.
 
-
 class Member(User):
     name = models.CharField(max_length=50)
-    phoneNumber = models.IntegerField()
+    phoneNumber = models.BigIntegerField()
     phoneNumberIsVerify = models.BooleanField(default=False)
     registerTime = models.DateTimeField(auto_now_add=True)
